Final commit
I have finished all of the scripts needed for the game and i've added a bit of my own twist to some of the map like some flying dragons and stuff. This commit will have the scripts as in where the hoop counts the points and goes darker with each point because it follows a color gradient. I can also come back to this commit to learn some more code that might come in handy in future projects.